More live migration bugfixes:
- Increase interval for logging live migration status, it's getting
  too noisy right now.

- Fix a logging error: do not clean up lock files for individual VMs
  as this may result in duplicate locks.

- Introduce a soft/hard layered timeout approach for pushing very
  high timeouts in specific situations to the incoming peer.

- Improve logging of timeout/wait cycles: log when the sleep time
  starts, not in the cycle after it.

- Ensure that we release the migration lock when doing another cycle
  of catching up with config changes.

- Remove the forced incoming migration service deregistration (cleanup)
  for now as the current version of consulate in production looks broken.

- Switch to a CSMA/CD-style backoff algorithm to avoid distributed
  lock contention under high migration pressure.

+++ src/fc/qemu/hazmat/qemu.py
@@ -365,7 +365,7 @@ class Qemu(object):
         for communicating status updates.
 
         """
-        timeout = TimeOut(timeout, 0.02, raise_on_timeout=True)
+        timeout = TimeOut(timeout, 1, raise_on_timeout=True)
         while timeout.tick():
             if timeout.interval < 10:
                 timeout.interval *= 1.4142
@@ -515,6 +515,11 @@ class Qemu(object):
             return
         self.log.debug('clean-run-files')
         for runfile in runfiles:
+            if runfile.endsdwith('.lock'):
+                # Never, ever, remove lock files. Those should be on
+                # partitions that get cleaned out during reboot, but
+                # never otherwise.
+                continue
             os.unlink(runfile)
 
     def prepare_config(self):
